fixing distortion issue

+++ backend.py
@@ -203,7 +203,7 @@ def test(path, tuning_system):
 
     # dynamic limiting or compression
     sr, x = wavfile.read(filepath)
-    x = float(x)
+    x = x.astype(np.float)
     x = x / np.abs(x).max() # x scale between -1 and 1
     x2 = dynamic_compression.limiter(x)
     x2 = np.int16(x2 * 32767)
